requirements.txt
Flask
gunicorn
aiohttp
python-dotenv